fix: zkstack readme update broken links and wording (#3375)
## What ❔

Hi @Deniallugo @popzxc 👋

Was playing with zkstack cli today. 

This PR is just fixing a couple of broken links and updating a bit of
wording in `zkstack_cli/README.md`.

PS: sidenote, "EVM emulator" is not explained in this readme, and
"Enable EVM emulator?" in interactive mode will be confusing to people
(they likely wont realize that it is emulating EVM equivalence but not
required for most chains + will be more expensive)

## Why ❔

Possibly a bit OCD, ie "Zk" being wrongly cased in one instance stayed
in my mind and bothered me hahahah 😉
<!-- Why are these changes done? What goal do they contribute to? What
are the principles behind them? -->
<!-- Example: PR templates ensure PR reviewers, observers, and future
iterators are in context about the evolution of repos. -->

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [x] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [x] Tests for the changes have been added / updated.
- [x] Documentation comments have been added / updated.
- [x] Code has been formatted via `zkstack dev fmt` and `zkstack dev
lint`.