feat(gateway): Migration to Gateway  (#3654)
## What ❔

Support migration to Gateway. That will be done through restart. 

There are few major changes in server

1. Contracats config now splitted to 3 configs: 
a. L1SpecificConfig - Stuff that exists only on l1 and will never be
migrated to gateway
  b. L2Contracts - Contracts that were deployed on l2
c. ChainSpecificContracts - Contracts that deployed on l1 and on gateway
2. L1 specific contracts and L2 contracts, for now, should be specified
in contracts config, manually. In a future we will be able to load them
as well
3. ChainSpecificContracts, are loading from the Settlement Layer, no
need to specify
4. There are no configuration for settlement layer. The settlement layer
known during the start of the node and the source of truth is contracts
both on l1 and settlement layer.
5. Server constantly checks that settlement layer has not changed. If it
chaged, server will wait until all processed transaction are sent and
confirmed and the necessary contracts are deployed to l1. After that
server will be killed. We assume that kubernetes will restart the server
6. During the start server checks the settlement layer and download the
necessary contracts config, later on almost all components of the system
are manipulating only with settlement layer client and config. Only eth
watcher knows about both layers.


External Node:
1. During the sync with the server External node verifies the
transaction on l1 for consistency. Once consistency checker is
panicking, external node crashes
2. During the start External node checks the latest eth txs in db and
depends on it setup the system to work either on Gateway or on L1



Migration process
1. Operator call the server notifier contract and this contract will
send the message to server, that all commit operations, should be
stopped
2. Server stops the commit operations, but continue to work in all other
parts
3. Operator call the migrate to gateway function on contracts
4. Once the migration is executed both on l1 and gateway, server will
crash
5. During the next start server will load all necessary contracts and
continue to work


## Why ❔

The migration to settlement layer should be smooth. Without necessary of
manual restarts and setting the contracts, it will drastically increase
the mistakes, especially with multiple chains in operation.




## Is this a breaking change?
- [ ] Yes
- [x] No

## Operational changes
<!-- Any config changes? Any new flags? Any changes to any scripts? -->
<!-- Please add anything that non-Matter Labs entities running their own
ZK Chain may need to know -->

## Checklist

<!-- Check your PR fulfills the following items. -->
<!-- For draft PRs check the boxes as you complete them. -->

- [x] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [x] Tests for the changes have been added / updated.
- [x] Documentation comments have been added / updated.
- [x] Code has been formatted via `zkstack dev fmt` and `zkstack dev
lint`.

---------

Signed-off-by: Danil <deniallugo@gmail.com>