feat(consensus): Validator committee rotation (#4014)
## What ❔

Adds the ability to change the validator committee from the consensus
registry, effectively allowing staking and similar features.

Initially the plan was to keep a list of the validator committees for
each epoch in the DAL and let the consensus component read it from
there, this started to be implemented in a past PR. Then the plan
changed to simplify this so the consensus component keeps that list of
committees instead (code in era-consensus) and just reads it directly
from the consensus registry contract.

We also deprecate some SQL tables that are no longer used/were never
used.

Fixes PLA-1112

## Is this a breaking change?
- [ ] Yes
- [X] No

## Operational changes

Instructions to use this feature will be added to the docs later.

## Checklist

- [x] PR title corresponds to the body of PR (we generate changelog
entries from PRs).
- [x] Tests for the changes have been added / updated.
- [x] Documentation comments have been added / updated.
- [x] Code has been formatted via `zkstack dev fmt` and `zkstack dev
lint`.

---------

Co-authored-by: Dima Zhornyk <55756184+dimazhornyk@users.noreply.github.com>